move view state to separate file
