+++ shared/src/commonMain/kotlin/org/hyperskill/app/progresses/presentation/ProgressScreenFeature.kt
@@ -1,7 +1,6 @@
 package org.hyperskill.app.progresses.presentation
 
 import org.hyperskill.app.analytic.domain.model.AnalyticEvent
-import org.hyperskill.app.projects.domain.model.ProjectLevel
 import org.hyperskill.app.projects.domain.model.ProjectWithProgress
 import org.hyperskill.app.track.domain.model.TrackWithProgress
 
@@ -14,49 +13,6 @@ object ProgressScreenFeature {
             val trackWithProgress: TrackWithProgress,
             val projectWithProgress: ProjectWithProgress
         ) : State
-    }
-
-    sealed interface ViewState {
-        object Idle : ViewState
-        object Loading : ViewState
-        object Error : ViewState
-        data class Content(
-            val trackProgress: TrackProgressViewState,
-            val projectProgress: ProjectProgressViewState?
-        ) : ViewState
-
-        data class TrackProgressViewState(
-            val title: String,
-
-            val imageSource: String?,
-
-            val completedTopicsCountLabel: String,
-            val completedTopicsPercentageLabel: String,
-            val completedTopicsPercentageProgress: Float,
-
-            val appliedTopicsCountLabel: String,
-            val appliedTopicsPercentageLabel: String,
-            val appliedTopicsPercentageProgress: Float,
-
-            val timeToCompleteLabel: String?,
-
-            val completedGraduateProjectsCount: Int,
-
-            val isCompleted: Boolean
-        )
-
-        data class ProjectProgressViewState(
-            val title: String,
-
-            val level: ProjectLevel?,
-
-            val timeToCompleteLabel: String?,
-
-            val completedStagesLabel: String,
-            val completedStagesProgress: Float,
-
-            val isCompleted: Boolean
-        )
     }
 
     sealed interface Message {
